// src/lwc/commerce_builder/pricingTiers/types.d.ts
export declare type PriceAdjustmentTier = {
    adjustmentValue?: string | null;
    id: string | null;
    lowerBound?: string | null;
    upperBound?: string | null;
    tierUnitPrice?: string | null;

    adjustmentValueFormat?: string | null;
};

export declare type SelectedVariant = {
    isValid?: boolean;
    options?: string[];
};